+++ src/handlers/proposalsHandlers.ts
@@ -8,6 +8,7 @@ import {
 import {
 	AuthenticatedRequest,
 	isId,
+	isAuthContext,
 	isTinyPgErrorWithQueryContext,
 	isWritableProposal,
 } from '../types';
@@ -17,6 +18,7 @@ import {
 	InputValidationError,
 } from '../errors';
 import {
+	extractCreatedByParameters,
 	extractOrganizationParameters,
 	extractPaginationParameters,
 	extractSearchParameters,
@@ -28,21 +30,25 @@ const getProposals = (
 	res: Response,
 	next: NextFunction,
 ): void => {
-	if (req.user === undefined) {
-		next(new FailedMiddlewareError('Unexpected lack of user context.'));
+	if (!isAuthContext(req)) {
+		next(new FailedMiddlewareError('Unexpected lack of auth context.'));
 		return;
 	}
-	const { user } = req;
 	const paginationParameters = extractPaginationParameters(req);
 	const searchParameters = extractSearchParameters(req);
 	const organizationParameters = extractOrganizationParameters(req);
+	const createdByParameters = extractCreatedByParameters(req);
+
 	(async () => {
-		const proposalBundle = await loadProposalBundle({
-			...getLimitValues(paginationParameters),
-			...searchParameters,
-			...organizationParameters,
-			createdBy: user.id,
-		});
+		const proposalBundle = await loadProposalBundle(
+			{
+				...getLimitValues(paginationParameters),
+				...searchParameters,
+				...organizationParameters,
+				...createdByParameters,
+			},
+			req,
+		);
 
 		res.status(200).contentType('application/json').send(proposalBundle);
 	})().catch((error: unknown) => {
@@ -87,8 +93,8 @@ const postProposal = (
 	res: Response,
 	next: NextFunction,
 ): void => {
-	if (req.user === undefined) {
-		next(new FailedMiddlewareError('Unexpected lack of user context.'));
+	if (!isAuthContext(req)) {
+		next(new FailedMiddlewareError('Unexpected lack of auth context.'));
 		return;
 	}
 	if (!isWritableProposal(req.body)) {