+++ src/types/index.ts
@@ -1,5 +1,6 @@
 export * from './ApplicationForm';
 export * from './ApplicationFormField';
+export * from './AuthContext';
 export * from './BaseField';
 export * from './BulkUpload';
 export * from './Bundle';